requirements.txt
numpy
scipy
scikit-image
opencv-python-headless
pyyaml
